Reorganized project structure for .NET Core
