+++ Epicoin.sln
@@ -1,25 +1,34 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.28010.2036
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epicoin", "Epicoin\Epicoin.csproj", "{C1F8F9CD-28E0-4AB6-A451-E09FA0BFB9BB}"
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Epicoin Core", "Core\Epicoin Core.csproj", "{D13E13C4-60AB-423B-80B6-A29E0FE6D606}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C1F8F9CD-28E0-4AB6-A451-E09FA0BFB9BB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C1F8F9CD-28E0-4AB6-A451-E09FA0BFB9BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C1F8F9CD-28E0-4AB6-A451-E09FA0BFB9BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C1F8F9CD-28E0-4AB6-A451-E09FA0BFB9BB}.Release|Any CPU.Build.0 = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {716B2264-4F9E-4759-84EB-CD0EF98C4135}
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|x64.Build.0 = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Debug|x86.Build.0 = Debug|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|x64.ActiveCfg = Release|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|x64.Build.0 = Release|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|x86.ActiveCfg = Release|Any CPU
+		{D13E13C4-60AB-423B-80B6-A29E0FE6D606}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
